requirements.txt
yt-dlp
numpy
openai-whisper
textsplit
sentence-transformers